moved widget components
