+++ src/widgets/blackScreen/blackScreen.css
