+++ src/widgets/calculator/calculator.css
