+++ src/widgets/calculator/calfFunc.js
