+++ src/widgets/covid/ChartDaily.jsx
