+++ src/widgets/covid/CovidChart.jsx
