+++ src/widgets/covid/chartDaily.css
