+++ src/widgets/covid/countries.js
