+++ src/widgets/covid/covid.css
