+++ src/widgets/covid/covidChart.css
