+++ src/widgets/forex/crypto.css
