+++ src/widgets/forex/forex.css
