+++ src/widgets/forex/forexData.js
