+++ src/widgets/news/eachNews.css
