+++ src/widgets/news/news.css
