+++ src/widgets/news/newsCategory.js
